Optimize day 15.
Because the squares all have the same distance from each other, we can
keep the open list of our breadth first search ordered and always use
the first destination, thus avoiding having to search for the
destination with the minimum distance.

By checking if we already reached a destination before adding it to the
open list we can drastically reduce the average length of the open list
and therefore speed up the search.